requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.4.1
attrs==21.2.0
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
constantly==15.1.0
cryptography==3.4.7
dateutils==0.6.12
distlib==0.3.2
Django==3.2.5
django-crispy-forms==1.13.0
django-social-share==2.2.1
djangorestframework==3.12.4
filelock==3.0.12
future==0.18.2
gevent==21.8.0
greenlet==1.1.2
html5lib==1.1
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
lxml==4.6.4
multitasking==0.0.10
nsetools==1.0.11
numpy==1.21.1
opencv-python==4.5.3.56
pandas==1.3.4
Pillow==8.3.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygame==2.0.1
PyPDF2==1.26.0
pypiwin32==223
pypng==0.0.20
PyQRCode==1.2.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.5
pyttsx3==2.7
pytube==11.0.1
pytz==2021.1
pywin32==301
razorpay==1.2.0
redis==3.5.3
reportlab==3.6.2
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
twisted-iocpsupport==1.0.2
typing-extensions==4.0.1
Unipath==1.1
urllib3==1.26.6
virtualenv==20.4.7
virtualenv-clone==0.5.6
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.2.2
whitenoise==5.3.0
xhtml2pdf==0.2.5
yfinance==0.1.67
zope.event==4.5.0
zope.interface==5.4.0
